New style validation should work... need to fix tests, improve error messages
