+++ src/ingest_validation_tools/directory-schemas/README.md
@@ -1,1 +1,1 @@
-These are JSON Schemas used by [directory-schema](https://github.com/hubmapconsortium/directory-schema/).
+TODO
