+++ src/ingest_validation_tools/directory-schemas/codex-akoya.yaml
@@ -1,49 +1,36 @@
-# A description of Akoya CODEX dataset directories.
+# A description of Stanford CODEX dataset directories.
 # Based on the description at https://docs.google.com/document/d/1CYYSXPQjwdbvmvZaEcsi_2udvDfGEZrMyh4yFnm4p3M/edit#heading=h.h6zxv8ghthzp
 
-items:
-  oneOf:
-    - $ref:
-        '#/definitions/subdirectory'
-    - properties:
-        type:
-          enum:
-            - file
-        name:
-          enum:
-            - experiment.json
-            - exposure_times.txt
-            - channelnames.txt
-            - segmentation.json
+# items:
+#   oneOf:
+#     - $ref:
+#         '#/definitions/subdirectory'
+#     - properties:
+#         type:
+#           enum:
+#             - file
+#         name:
+#           enum:
+#             - channelNames.txt
+#             - Experiment.json
+#             - exposure_times.txt
+#             - processingOptions.json
 
-contains:
-  properties:
-      name:
-        enum:
-          - experiment.json
-  properties:
-      name:
-        enum:
-          - exposure_times.txt
-  properties:
-      name:
-        enum:
-          - channelNames.txt
-  properties:
-      name:
-        enum:
-          - segmentation.json
+channelNames.txt: 1
+Experiment.json: 1
+exposure_times.txt: 1
+processingOptions.json: 1
 
-definitions:
-  subdirectory:
-    properties:
-      type:
-        enum:
-          - directory
-      name:
-        pattern: '^cyc\d+_reg\d+_\d+_\d+$'
-      contents:
-        items:
-          properties:
-            name:
-              pattern: '^\d+_\d+_Z\d+_CH\d+\.tif|\d+\.gci$'
+# definitions:
+#   subdirectory:
+#     properties:
+#       type:
+#         enum:
+#           - directory
+#       name:
+#         pattern: '^Cyc\d+_reg\d+$'
+#       contents:
+#         items:
+#           properties:
+#             name:
+#               pattern: '^HE_\d+_Z\d+_CH\d+$'
