+++ src/ingest_validation_tools/directory-schemas/imc.yaml
@@ -1,1 +1,2 @@
-items: {} # TODO
+'**': 0
+# TODO
