+++ src/ingest_validation_tools/directory-schemas/lcms.yaml
@@ -1,0 +1,2 @@
+'**': 0
+# TODO
