+++ src/ingest_validation_tools/directory-schemas/mixif.yaml
@@ -1,2 +1,2 @@
-items: {}
-# TODO: Flesh out
+'**': 0
+# TODO
