+++ src/ingest_validation_tools/directory-schemas/scatacseq.yaml
@@ -1,7 +1,2 @@
-items:
-  properties:
-    type:
-      enum:
-        - file
-    name:
-      pattern: '\.fastq(\.gz)?$'
+'*.fastq': 0
+'*.fastq.gz': 0
