+++ src/ingest_validation_tools/directory-schemas/wgs.yaml
@@ -1,1 +1,2 @@
-items: {} # TODO
+'**': 0
+# TODO
